+++ libraries/base/GHC/Base.hs
@@ -1162,7 +1162,6 @@ The rules for map work like this.
 
 {-# RULES "map/coerce" [1] map coerce = coerce #-}
 
-
 ----------------------------------------------
 --              append
 ----------------------------------------------